KAN-107 refactor: hook import문 정리
