+++ src/app/hooks/index.ts
@@ -13,3 +13,10 @@ export { useRefresh } from './auth/useRefresh'
 export { useChangePassword } from './auth/useChangePassword'
 export { useCompleteJoin } from './auth/useCompleteJoin'
 export { useFindPassword } from './auth/useFindPassword'
+
+// profile
+export { useGetMeProfile } from './profile/useProfileQueries'
+
+// products
+export { useProducts } from './products/useProductsMutation'
+export { useGetProductDetail, useGetProductList } from './products/useProductsQueries'
